backend/requirements.txt
# FastAPI and web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0
httpx==0.26.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Redis and Celery
redis==5.0.1
celery[redis]==5.3.6

# S3/MinIO
boto3==1.34.25
aioboto3==12.3.0

# Validation and serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# AI/ML (will be used by services)
openai==1.10.0
google-generativeai==0.3.2
tiktoken==0.5.2

# Audio processing
numpy==1.26.3
scipy==1.12.0
soundfile==0.12.1
librosa==0.10.1

# Image/Video processing
pillow==10.2.0
opencv-python-headless==4.9.0.80

# Utilities
python-dotenv==1.0.0
structlog==24.1.0
tenacity==8.2.3
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
isort==5.13.2
mypy==1.8.0

